+++ src/Flags/index.js
@@ -1,3 +1,0 @@
-export { default as DataFlag } from "./DataFlag";
-export { default as DataFlags } from "./DataFlags";
-export { default as useDataFlags } from "./useDataFlags";