+++ CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.2.2] - 2022-03-01
+
+- fix: save the data download to disk instead of linking
+- fix: do not include monthly totals for the current month
+- fix: move events legend above summary
+- content: update about intro
+
+## [3.2.1] - 2022-02-08
+
+- fix: no longer crashes when manually typing date
+
+## [3.2.0] - 2022-02-07
+
+- feat: show events on the time series
+- feat: add download data button
+- fix: adjust time series styles so 0 is always visible
+- fix: parse dates consistently (with `parseDate` util)
+
 ## [3.1.1] - 2022-02-01
 
 - fix: update fillSeries util to be inclusive of end date (fixes monthly totals)